replace Nav component
