+++ pages/settings/calorie-goal.js
@@ -1,7 +1,7 @@
 import { useEffect, useState  } from 'react';
 import Head from 'next/head';
 
-import SettingsNav from '../../shared/components/SettingsNav';
+import Nav from '../../shared/components/Nav';
 import Snackbar from '../../shared/components/Snackbar/Snackbar';
 
 import useSnackbar from '../../shared/hooks/useSnackbar';
@@ -46,7 +46,7 @@ export default function CalorieGoal () {
             </Head>
 
             <div className="container">
-                <SettingsNav />
+                <Nav />
 
                 <h1 className="mt-1 mx-15">
                     calorie goal
