+++ pages/settings/data/clear-data.js
@@ -1,6 +1,6 @@
 import Head from 'next/head';
 
-import SettingsNav from '../../../shared/components/SettingsNav';
+import Nav from '../../../shared/components/Nav';
 import Snackbar from '../../../shared/components/Snackbar/Snackbar';
 
 import useSnackbar from '../../../shared/hooks/useSnackbar';
@@ -31,7 +31,8 @@ export default function ClearData () {
             </Head>
 
             <div className="container">
-                <SettingsNav />
+                <Nav />
+
                 <div className="mx-15">
                     <h1 className="mt-1">
                         clear data
