+++ pages/settings/data/import-data.js
@@ -1,6 +1,6 @@
 import Head from 'next/head';
 
-import SettingsNav from '../../../shared/components/SettingsNav';
+import Nav from '../../../shared/components/Nav';
 import Snackbar from '../../../shared/components/Snackbar/Snackbar';
 
 import { importFoodBlocks, importFoodDictionary } from '../../../shared/food/food';
@@ -64,7 +64,8 @@ export default function ImportData () {
                 <link rel="icon" href="/favicon.ico" />
             </Head>
             <div className="container">
-                <SettingsNav />
+                <Nav />
+
                 <div className="mx-15">
                     <h1 className="mt-1">
                         import data
