+++ pages/settings/data/index.js
@@ -1,6 +1,6 @@
 import Head from 'next/head';
 
-import SettingsNav from '../../../shared/components/SettingsNav';
+import Nav from '../../../shared/components/Nav';
 import SettingsItem from '../../../shared/components/SettingsItems/SettingsItem';
 
 export default function Data () {
@@ -13,7 +13,7 @@ export default function Data () {
             </Head>
 
             <div className="container">
-                <SettingsNav />
+                <Nav />
                 <div className="mx-15">
                     <h1 className="mt-1">
                         data
