+++ pages/settings/index.js
@@ -1,7 +1,7 @@
 import Head from 'next/head';
 
-import SettingsNav from '../../shared/components/SettingsNav';
 import SettingsItem from '../../shared/components/SettingsItems/SettingsItem';
+import Nav from '../../shared/components/Nav';
 
 export default function Settings() {
     return (
@@ -13,7 +13,8 @@ export default function Settings() {
             </Head>
 
             <div className="container">
-                <SettingsNav />
+                <Nav />
+
                 <div className="mx-15">
                     <h1 className="mt-1">
                         settings
