+++ shared/components/Nav.js
@@ -1,0 +1,70 @@
+import Link from 'next/link';
+import { useRouter } from 'next/router';
+import ThemeSelector from './ThemeSelector';
+
+export default function Nav ({ link=null }) {
+    const router = useRouter();
+    const date = router.query.date;
+
+    return (
+        <nav className="nav pt-1">
+            <Link href="/">
+                <a class="icon mb-1" title="calendar">
+                    <svg
+                        xmlns="http://www.w3.org/2000/svg"
+                        viewBox="0 0 24 24"
+                        width="24"
+                        height="24">
+                        <path fill="none" d="M0 0h24v24H0z"/>
+                        <path d="M17 3h4a1 1 0 0 1 1 1v16a1 1 0 0 1-1 1H3a1 1 0 0 1-1-1V4a1 1 0 0 1 1-1h4V1h2v2h6V1h2v2zm-2 2H9v2H7V5H4v4h16V5h-3v2h-2V5zm5 6H4v8h16v-8z"/>
+                    </svg>
+                </a>
+            </Link>
+
+            { link ? (
+                <Link href={ link.link }>
+                    <a className="icon mb-1" title={ link.text }>
+                        <svg
+                            xmlns="http://www.w3.org/2000/svg"
+                            viewBox="0 0 24 24"
+                            width="24"
+                            height="24">
+                            <path fill="none" d="M0 0h24v24H0z"/>
+                            <path d="M17 3h4a1 1 0 0 1 1 1v16a1 1 0 0 1-1 1H3a1 1 0 0 1-1-1V4a1 1 0 0 1 1-1h4V1h2v2h6V1h2v2zm3 6V5h-3v2h-2V5H9v2H7V5H4v4h16zm0 2H4v8h16v-8zM6 13h5v4H6v-4z"/>
+                        </svg> 
+                    </a>
+                </Link>
+            ) : (null) }
+
+            { link ? (
+                <Link href={`/day/${ date }/blocks`}>
+                    <a className="icon mb-1" title="blocks">
+                        <svg
+                            xmlns="http://www.w3.org/2000/svg"
+                            viewBox="0 0 24 24"
+                            width="24"
+                            height="24">
+                            <path fill="none" d="M0 0h24v24H0z"/>
+                            <path d="M11 4h10v2H11V4zm0 4h6v2h-6V8zm0 6h10v2H11v-2zm0 4h6v2h-6v-2zM3 4h6v6H3V4zm2 2v2h2V6H5zm-2 8h6v6H3v-6zm2 2v2h2v-2H5z"/>
+                        </svg>
+                    </a>
+                </Link>
+            ) : (null) }
+
+            <Link href="/settings">
+                <a className="icon mb-1" title="settings">
+                    <svg
+                        xmlns="http://www.w3.org/2000/svg"
+                        viewBox="0 0 24 24"
+                        width="24"
+                        height="24">
+                        <path fill="none" d="M0 0h24v24H0z"/>
+                        <path d="M2.132 13.63a9.942 9.942 0 0 1 0-3.26c1.102.026 2.092-.502 2.477-1.431.385-.93.058-2.004-.74-2.763a9.942 9.942 0 0 1 2.306-2.307c.76.798 1.834 1.125 2.764.74.93-.385 1.457-1.376 1.43-2.477a9.942 9.942 0 0 1 3.262 0c-.027 1.102.501 2.092 1.43 2.477.93.385 2.004.058 2.763-.74a9.942 9.942 0 0 1 2.307 2.306c-.798.76-1.125 1.834-.74 2.764.385.93 1.376 1.457 2.477 1.43a9.942 9.942 0 0 1 0 3.262c-1.102-.027-2.092.501-2.477 1.43-.385.93-.058 2.004.74 2.763a9.942 9.942 0 0 1-2.306 2.307c-.76-.798-1.834-1.125-2.764-.74-.93.385-1.457 1.376-1.43 2.477a9.942 9.942 0 0 1-3.262 0c.027-1.102-.501-2.092-1.43-2.477-.93-.385-2.004-.058-2.763.74a9.942 9.942 0 0 1-2.307-2.306c.798-.76 1.125-1.834.74-2.764-.385-.93-1.376-1.457-2.477-1.43zM12 15a3 3 0 1 0 0-6 3 3 0 0 0 0 6z"/>
+                    </svg>
+                </a>
+            </Link>
+
+            <ThemeSelector />
+        </nav>
+    );
+}
