+++ styles/module.scss
@@ -77,8 +77,7 @@
 }
 
 .container {
-  margin: 0 auto;
-  max-width: 1600px;
+  margin-left: 40px;
 }
 
 .modal {
@@ -210,7 +209,7 @@ body.modal-is-open {
   background-color: #030852 !important;
 }
 
-.theme-selector {
+.icon {
   border-radius: 50%;
   display: flex;
   align-items: center;
@@ -219,4 +218,13 @@ body.modal-is-open {
   width: 36px;
 
   background-color: #ffffff;
+}
+
+.nav {
+  position: fixed;
+  top: 0;
+  left: 0;
+  height: 100%;
+  width: 40px;
+  background-color: #C75000;
 }
