+++ styles/utilities.scss
@@ -46,12 +46,12 @@
   margin-bottom: 20px;
 }
 
-.pt-1 {
-  padding-top: 10px;
+.ml-1 {
+  margin-left: 10px;
 }
 
-.pt-3 {
-  padding-top: 30px;
+.pt-1 {
+  padding-top: 10px;
 }
 
 .p-1 {
